requirements.txt
Django==3.2
django-environ==0.9.0
pre-commit==2.19.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
